+++ nodes/Autotask/operations/common/get-many/description.ts
@@ -8,7 +8,7 @@ export const getManyOptions: INodeProperties[] = [
 		displayName: 'Get All',
 		name: 'returnAll',
 		type: 'boolean',
-		default: true,
+		default: false,
 		description: 'Whether to return all results or only up to a given limit',
 		displayOptions: {
 			show: {