+++ nodes/Autotask/resources/serviceCalls/execute.ts
@@ -5,14 +5,15 @@ import {
 	UpdateOperation,
 	GetOperation,
 	GetManyOperation,
+	DeleteOperation,
 	CountOperation,
 } from '../../operations/base';
 import { executeEntityInfoOperations } from '../../operations/common/entityInfo.execute';
 import { handleGetManyAdvancedOperation } from '../../operations/common/get-many-advanced';
 
-const ENTITY_TYPE = 'contact';
+const ENTITY_TYPE = 'servicecall';
 
-export async function executeContactOperation(
+export async function executeServiceCallOperation(
 	this: IExecuteFunctions,
 ): Promise<INodeExecutionData[][]> {
 	const items = this.getInputData();
@@ -30,9 +31,9 @@ export async function executeContactOperation(
 				}
 
 				case 'update': {
-					const entityId = this.getNodeParameter('id', i) as string;
+					const serviceCallId = this.getNodeParameter('id', i) as string;
 					const updateOp = new UpdateOperation<IAutotaskEntity>(ENTITY_TYPE, this);
-					const response = await updateOp.execute(i, entityId);
+					const response = await updateOp.execute(i, serviceCallId);
 					returnData.push({ json: response });
 					break;
 				}
@@ -52,9 +53,10 @@ export async function executeContactOperation(
 					break;
 				}
 
-				case 'getManyAdvanced': {
-					const results = await handleGetManyAdvancedOperation.call(this, ENTITY_TYPE, i);
-					returnData.push(...results);
+				case 'delete': {
+					const deleteOp = new DeleteOperation<IAutotaskEntity>(ENTITY_TYPE, this);
+					await deleteOp.execute(i);
+					returnData.push({ json: { success: true } });
 					break;
 				}
 
@@ -67,6 +69,12 @@ export async function executeContactOperation(
 							entityType: ENTITY_TYPE,
 						},
 					});
+					break;
+				}
+
+				case 'getManyAdvanced': {
+					const results = await handleGetManyAdvancedOperation.call(this, ENTITY_TYPE, i);
+					returnData.push(...results);
 					break;
 				}